+++ app/components/trackItem.js
@@ -2,14 +2,17 @@ import { Pressable, StyleSheet, Text, View, Image } from "react-native";
 import { millisToMinutesAndSeconds } from "../../utils";
 import { Themes } from "../../assets/Themes";
 import { Ionicons } from "@expo/vector-icons";
+import { NavigationContainer } from "@react-navigation/native";
 
 export default function TrackItem({
-  index,
   imageUrl,
   songTitle,
   artist,
   album,
   duration,
+  preview_url,
+  external_url,
+  navigation,
 }) {
   return (
     <View style={[styles.item]}>
@@ -21,7 +24,11 @@ export default function TrackItem({
           marginRight: "2%",
         }}
       >
-        <Pressable onPress={(event) => handleClick()}>
+        <Pressable
+          onPress={() =>
+            navigation.navigate("Song preview", { url: preview_url })
+          }
+        >
           <Ionicons
             name="play-circle"
             size={20}
@@ -33,43 +40,51 @@ export default function TrackItem({
           />
         </Pressable>
       </View>
-      <View style={{ borderColor: "green", width: "25%" }}>
-        <Image
-          style={{
-            width: 70,
-            height: 70,
-            marginLeft: "5%",
-            marginBottom: "15%",
-            justifyContent: "center",
-            resizeMode: "contain",
-          }}
-          source={{ uri: imageUrl }}
-        />
-      </View>
-      <View
-        style={{ width: "35%", flexDirection: "column", paddingRight: "2%" }}
+
+      <Pressable
+        onPress={() =>
+          navigation.navigate("Song details", { url: external_url })
+        }
+        style={{ flexDirection: "row", justifyContent: "space-between" }}
       >
-        <Text numberOfLines={1} style={styles.textSection}>
-          {songTitle}
-        </Text>
-        <Text
-          numberOfLines={1}
-          style={[styles.textSection, { color: Themes.colors.gray }]}
+        <View style={{ borderColor: "green", width: "25%" }}>
+          <Image
+            style={{
+              width: 70,
+              height: 70,
+              marginLeft: "5%",
+              marginBottom: "15%",
+              justifyContent: "center",
+              resizeMode: "contain",
+            }}
+            source={{ uri: imageUrl }}
+          />
+        </View>
+        <View
+          style={{ width: "35%", flexDirection: "column", paddingRight: "2%" }}
         >
-          {artist}
-        </Text>
-      </View>
-      <View style={{ width: "22%" }}>
-        <Text numberOfLines={1} style={styles.textSection}>
-          {album}
-        </Text>
-      </View>
-      <View>
-        <Text numberOfLines={1} style={styles.textSection}>
-          {" "}
-          {millisToMinutesAndSeconds(duration)}{" "}
-        </Text>
-      </View>
+          <Text numberOfLines={1} style={styles.textSection}>
+            {songTitle}
+          </Text>
+          <Text
+            numberOfLines={1}
+            style={[styles.textSection, { color: Themes.colors.gray }]}
+          >
+            {artist}
+          </Text>
+        </View>
+        <View style={{ width: "22%" }}>
+          <Text numberOfLines={1} style={styles.textSection}>
+            {album}
+          </Text>
+        </View>
+        <View>
+          <Text numberOfLines={1} style={styles.textSection}>
+            {" "}
+            {millisToMinutesAndSeconds(duration)}{" "}
+          </Text>
+        </View>
+      </Pressable>
     </View>
   );
 }
@@ -78,7 +93,6 @@ const styles = StyleSheet.create({
   item: {
     flexDirection: "row",
     width: "100%",
-    justifyContent: "space-between",
   },
   textSection: {
     color: Themes.colors.white,
