+++ package.json
@@ -21,6 +21,8 @@
     ]
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.13",
+    "@react-navigation/stack": "^6.3.4",
     "axios": "^0.24.0",
     "caniuse-lite": "^1.0.30001338",
     "expo": "~44.0.0",
@@ -33,7 +35,11 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-app-auth": "^6.4.0",
+    "react-native-gesture-handler": "~2.1.0",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "~3.10.1",
     "react-native-web": "0.17.1",
+    "react-native-webview": "11.15.0",
     "react-test-renderer": "~17.0.0"
   },
   "devDependencies": {
